tools: pkgworker: refactor as a lib instead of a main package
