+++ toolkit/tools/pkg/pkgworker/pkgworker.go
@@ -1,0 +1,354 @@
+package pkgworker
+
+import (
+	"fmt"
+	"os"
+	"path/filepath"
+	"regexp"
+	"strings"
+	"time"
+
+	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/file"
+	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/logger"
+	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/packagerepo/repomanager/rpmrepomanager"
+	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/rpm"
+	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/safechroot"
+	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/shell"
+	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/sliceutils"
+)
+
+const (
+	chrootRpmBuildRoot      = "/usr/src/mariner"
+	chrootLocalRpmsDir      = "/localrpms"
+	chrootLocalRpmsCacheDir = "/upstream-cached-rpms"
+)
+
+var (
+	brPackageNameRegex        = regexp.MustCompile(`^[^\s]+`)
+	equalToRegex              = regexp.MustCompile(` '?='? `)
+	greaterThanOrEqualRegex   = regexp.MustCompile(` '?>='? [^ ]*`)
+	installedPackageNameRegex = regexp.MustCompile(`^(.+)(-[^-]+-[^-]+)`)
+	lessThanOrEqualToRegex    = regexp.MustCompile(` '?<='? `)
+	packageUnavailableRegex   = regexp.MustCompile(`^No package \\x1b\[1m\\x1b\[30m(.+) \\x1b\[0mavailable`)
+)
+
+func (cfg *Config) BuildSRPMInChrootAndCopyToOut() ([]string, error) {
+	rpmsDirAbsPath, err := filepath.Abs(cfg.RpmsDirPath)
+	// Todo .. do not panic here .. but bubble it up accordingly
+	logger.PanicOnError(err, "Unable to find absolute path for RPMs directory '%s'", cfg.RpmsDirPath)
+
+	srpmsDirAbsPath, err := filepath.Abs(cfg.SrpmsDirPath)
+	logger.PanicOnError(err, "Unable to find absolute path for SRPMs directory '%s'", cfg.SrpmsDirPath)
+
+	srpmName := strings.TrimSuffix(filepath.Base(cfg.SrpmFile), ".src.rpm")
+	chrootDir := filepath.Join(cfg.WorkDir, srpmName)
+
+	defines := rpm.DefaultDefines(cfg.RunCheck)
+	defines[rpm.DistTagDefine] = cfg.DistTag
+	defines[rpm.DistroReleaseVersionDefine] = cfg.DistroReleaseVersion
+	defines[rpm.DistroBuildNumberDefine] = cfg.DistroBuildNumber
+
+	builtRPMs, err := buildSRPMInChroot(chrootDir, rpmsDirAbsPath, cfg.WorkerTar, cfg.SrpmFile, cfg.RepoFile, cfg.RpmmacrosFile, cfg.CacheDir, defines, cfg.NoCleanup, cfg.RunCheck, cfg.PackagesToInstall)
+	err = copySRPMToOutput(cfg.SrpmFile, srpmsDirAbsPath)
+	logger.PanicOnError(err, "Failed to copy SRPM '%s' to output directory '%s'.", cfg.SrpmFile, rpmsDirAbsPath)
+
+	return builtRPMs, err
+}
+
+func copySRPMToOutput(srpmFilePath, srpmOutputDirPath string) (err error) {
+	const srpmsDirName = "SRPMS"
+
+	srpmFileName := filepath.Base(srpmFilePath)
+	srpmOutputFilePath := filepath.Join(srpmOutputDirPath, srpmFileName)
+
+	err = file.Copy(srpmFilePath, srpmOutputFilePath)
+
+	return
+}
+
+func buildSRPMInChroot(chrootDir, rpmDirPath, workerTar, srpmFile, repoFile, rpmmacrosFile, cacheDir string, defines map[string]string, noCleanup, runCheck bool, packagesToInstall []string) (builtRPMs []string, err error) {
+	const (
+		buildHeartbeatTimeout = 30 * time.Minute
+
+		existingChrootDir = false
+		squashErrors      = false
+
+		overlaySource  = ""
+		overlayWorkDir = "/overlaywork"
+		rpmDirName     = "RPMS"
+	)
+
+	srpmBaseName := filepath.Base(srpmFile)
+
+	quit := make(chan bool)
+	go func() {
+		logger.Log.Infof("Building (%s).", srpmBaseName)
+
+		for {
+			select {
+			case <-quit:
+				if err == nil {
+					logger.Log.Infof("Built (%s) -> %v.", srpmBaseName, builtRPMs)
+				}
+				return
+			case <-time.After(buildHeartbeatTimeout):
+				logger.Log.Infof("Heartbeat: still building (%s).", srpmBaseName)
+			}
+		}
+	}()
+	defer func() {
+		quit <- true
+	}()
+
+	// Create the chroot used to build the SRPM
+	chroot := safechroot.NewChroot(chrootDir, existingChrootDir)
+
+	overlayMount, overlayExtraDirs := safechroot.NewOverlayMountPoint(chroot.RootDir(), overlaySource, chrootLocalRpmsDir, rpmDirPath, chrootLocalRpmsDir, overlayWorkDir)
+	rpmCacheMount := safechroot.NewMountPoint(cacheDir, chrootLocalRpmsCacheDir, "", safechroot.BindMountPointFlags, "")
+	mountPoints := []*safechroot.MountPoint{overlayMount, rpmCacheMount}
+	extraDirs := append(overlayExtraDirs, chrootLocalRpmsCacheDir)
+
+	err = chroot.Initialize(workerTar, extraDirs, mountPoints)
+	if err != nil {
+		return
+	}
+	defer chroot.Close(noCleanup)
+
+	// Place extra files that will be needed to build into the chroot
+	srpmFileInChroot, err := copyFilesIntoChroot(chroot, srpmFile, repoFile, rpmmacrosFile, runCheck)
+	if err != nil {
+		return
+	}
+
+	err = chroot.Run(func() (err error) {
+		return buildRPMFromSRPMInChroot(srpmFileInChroot, runCheck, defines, packagesToInstall)
+	})
+	if err != nil {
+		return
+	}
+
+	rpmBuildOutputDir := filepath.Join(chroot.RootDir(), chrootRpmBuildRoot, rpmDirName)
+	builtRPMs, err = moveBuiltRPMs(rpmBuildOutputDir, rpmDirPath)
+
+	return
+}
+
+func buildRPMFromSRPMInChroot(srpmFile string, runCheck bool, defines map[string]string, packagesToInstall []string) (err error) {
+	// Convert /localrpms into a repository that a package manager can use.
+	err = rpmrepomanager.CreateRepo(chrootLocalRpmsDir)
+	if err != nil {
+		return
+	}
+
+	// install any additional packages, such as build dependencies.
+	err = tdnfInstall(packagesToInstall)
+	if err != nil {
+		return
+	}
+
+	// Remove all libarchive files on the system before issuing a build.
+	// If the build environment has libtool archive files present, gnu configure
+	// could detect it and create more libtool archive files which can cause
+	// build failures.
+	err = removeLibArchivesFromSystem()
+	if err != nil {
+		return
+	}
+
+	// Build the SRPM
+	if runCheck {
+		err = rpm.BuildRPMFromSRPM(srpmFile, defines)
+	} else {
+		err = rpm.BuildRPMFromSRPM(srpmFile, defines, "--nocheck")
+	}
+
+	return
+}
+
+func moveBuiltRPMs(rpmOutDir, dstDir string) (builtRPMs []string, err error) {
+	const rpmExtension = ".rpm"
+	err = filepath.Walk(rpmOutDir, func(path string, info os.FileInfo, fileErr error) (err error) {
+		if fileErr != nil {
+			return fileErr
+		}
+
+		// Only copy regular files (not unix sockets, directories, links, ...)
+		if !info.Mode().IsRegular() {
+			return
+		}
+
+		if !strings.HasSuffix(path, rpmExtension) {
+			return
+		}
+
+		// Get the relative path of the RPM, this will include the architecture directory it lives in.
+		// Then join the relative path to the destination directory, this will ensure the RPM gets placed
+		// in its correct architecture directory.
+		relPath, err := filepath.Rel(rpmOutDir, path)
+		if err != nil {
+			return
+		}
+
+		dstFile := filepath.Join(dstDir, relPath)
+		err = file.Move(path, dstFile)
+		if err != nil {
+			return
+		}
+
+		builtRPMs = append(builtRPMs, dstFile)
+		return
+	})
+
+	return
+}
+
+func tdnfInstall(packages []string) (err error) {
+	const (
+		alreadyInstalledPostfix = "is already installed"
+		noMatchingPackagesErr   = "Error(1011) : No matching packages"
+		packageMatchGroup       = 1
+	)
+
+	if len(packages) == 0 {
+		return
+	}
+
+	// TDNF supports requesting versioned packages in the form of {name}-{version}.{dist}.{arch}.
+	// The packages to install list may contain file paths to rpm files so those will need to be filtered:
+	// - Strip any .rpm from packages as TDNF does not support requesting a package with the extension.
+	// - Strip any filepath from packages.
+	for i := range packages {
+		packages[i] = filepath.Base(strings.TrimSuffix(packages[i], ".rpm"))
+	}
+
+	installArgs := []string{"install", "-y"}
+	installArgs = append(installArgs, packages...)
+	stdout, stderr, err := shell.Execute("tdnf", installArgs...)
+	foundNoMatchingPackages := false
+
+	if err != nil {
+		logger.Log.Warnf("Failed to install build requirements. stderr: %s\nstdout: %s", stderr, stdout)
+		// TDNF will output an error if all packages are already installed.
+		// Ignore it iff there is no other error present in stderr.
+		splitStderr := strings.Split(stderr, "\n")
+		for _, line := range splitStderr {
+			trimmedLine := strings.TrimSpace(line)
+			if trimmedLine == "" {
+				continue
+			}
+
+			if strings.Contains(trimmedLine, noMatchingPackagesErr) {
+				foundNoMatchingPackages = true
+			}
+
+			if !strings.HasSuffix(trimmedLine, alreadyInstalledPostfix) && trimmedLine != noMatchingPackagesErr {
+				err = fmt.Errorf(trimmedLine)
+				return
+			}
+		}
+		err = nil
+	}
+
+	// TDNF will ignore unavailable packages that have been requested to be installed without reporting an error code.
+	// Search the stdout of TDNF for such a failure and warn the user.
+	// This may happen if a SPEC requires the the path to a tool (e.g. /bin/cp), so mark it as a warning for now.
+	var failedToInstall []string
+	splitStdout := strings.Split(stdout, "\n")
+	for _, line := range splitStdout {
+		trimmedLine := strings.TrimSpace(line)
+		matches := packageUnavailableRegex.FindStringSubmatch(trimmedLine)
+		if len(matches) == 0 {
+			continue
+		}
+
+		failedToInstall = append(failedToInstall, matches[packageMatchGroup])
+	}
+
+	// TDNF will output the error "Error(1011) : No matching packages" if all packages could not be found.
+	// In this case it will not print any of the individual packages that failed.
+	if foundNoMatchingPackages && len(failedToInstall) == 0 {
+		failedToInstall = packages
+	}
+
+	if len(failedToInstall) != 0 {
+		err = fmt.Errorf("unable to install the following packages: %v", failedToInstall)
+	}
+
+	return
+}
+
+// removeLibArchivesFromSystem removes all libarchive files on the system. If
+// the build environment has libtool archive files present, gnu configure could
+// detect it and create more libtool archive files which can cause build failures.
+func removeLibArchivesFromSystem() (err error) {
+	dirsToExclude := []string{"/proc", "/dev", "/sys", "/run"}
+
+	err = filepath.Walk("/", func(path string, info os.FileInfo, err error) error {
+		if err != nil {
+			return err
+		}
+
+		// Skip directories that are meant for device files and kernel virtual filesystems.
+		// These will not contain .la files and are mounted into the safechroot from the host.
+		if info.IsDir() && sliceutils.Contains(dirsToExclude, path, sliceutils.StringMatch) {
+			return filepath.SkipDir
+		}
+
+		if strings.HasSuffix(info.Name(), ".la") {
+			return os.Remove(path)
+		}
+
+		return nil
+	})
+
+	if err != nil {
+		logger.Log.Warnf("Unable to remove lib archive file: %s", err)
+	}
+
+	return
+}
+
+// copyFilesIntoChroot copies several required build specific files into the chroot.
+func copyFilesIntoChroot(chroot *safechroot.Chroot, srpmFile, repoFile, rpmmacrosFile string, runCheck bool) (srpmFileInChroot string, err error) {
+	const (
+		chrootRepoDestDir = "/etc/yum.repos.d"
+		chrootSrpmDestDir = "/root/SRPMS"
+		resolvFilePath    = "/etc/resolv.conf"
+		rpmmacrosDest     = "/usr/lib/rpm/macros.d/macros.override"
+	)
+
+	repoFileInChroot := filepath.Join(chrootRepoDestDir, filepath.Base(repoFile))
+	srpmFileInChroot = filepath.Join(chrootSrpmDestDir, filepath.Base(srpmFile))
+
+	filesToCopy := []safechroot.FileToCopy{
+		safechroot.FileToCopy{
+			Src:  repoFile,
+			Dest: repoFileInChroot,
+		},
+		safechroot.FileToCopy{
+			Src:  srpmFile,
+			Dest: srpmFileInChroot,
+		},
+	}
+
+	if rpmmacrosFile != "" {
+		rpmmacrosCopy := safechroot.FileToCopy{
+			Src:  rpmmacrosFile,
+			Dest: rpmmacrosDest,
+		}
+		filesToCopy = append(filesToCopy, rpmmacrosCopy)
+	}
+
+	if runCheck {
+		logger.Log.Debug("Enabling network access because we're running package tests.")
+
+		resolvFileCopy := safechroot.FileToCopy{
+			Src:  resolvFilePath,
+			Dest: resolvFilePath,
+		}
+		filesToCopy = append(filesToCopy, resolvFileCopy)
+	}
+
+	err = chroot.AddFiles(filesToCopy...)
+	return
+}
